fix: code review 반영

+++ src/product/application/commands/delete-product/delete-product.handler.ts
@@ -1,6 +1,6 @@
 import { CommandHandler } from '@nestjs/cqrs';
 import { DeleteProductCommand } from './delete-product.command';
-import { Inject } from '@nestjs/common';
+import { Inject, NotFoundException } from '@nestjs/common';
 import {
   PRODUCT_REPOSITORY,
   ProductRepository,
@@ -14,6 +14,11 @@ export class DeleteProductHandler {
   ) {}
   async execute(command: DeleteProductCommand) {
     const product = await this.repository.findById(command.id);
+    if (!product) {
+      throw new NotFoundException(
+        `ID가 ${command.id}인 제품을 찾을 수 없습니다.`,
+      );
+    }
     product.delete();
     await this.repository.delete(product);
   }
